Inspect and Update alternative matches for a name returned by tnrs_match_names

Taxonomic names may have different meanings in different taxonomic contexts, as the same genus name can be applied to animals and plants for instance. Additionally, the meaning of a taxonomic name may have change throughout its history, and may have referred to a different taxon in the past. In such cases, a given names might have multiple matches in the Open Tree Taxonomy. These functions allow users to inspect (and update) alternative meaning of a given name and its current taxonomic status according to the Open Tree Taxonomy.

# S3 method for match_names
inspect(response, row_number, taxon_name, ott_id,
  ...)

inspect(response, ...)

# S3 method for match_names
update(object, row_number, taxon_name, ott_id,
  new_row_number, new_ott_id, ...)

Arguments

response
an object generated by the tnrs_match_names function

row_number
the row number corresponding to the name to inspect

taxon_name
the taxon name corresponding to the name to inspect

ott_id
the ott id corresponding to the name to inspect

...
currently ignored

object
an object created by tnrs_match_names

new_row_number
the row number in the output of inspect to replace the taxa specified by row_number, taxon_name, or ott_id.

new_ott_id
the ott id of the taxon to replace the taxa specified by row_number, taxon_name, or ott_id.

Value

a data frame

Details

To inspect alternative taxonomic meanings of a given name, you need to provide the object resulting from a call to the tnrs_match_names function, as well as one of either the row number corresponding to the name in this object, the name itself (as used in the original query), or the ott_id listed for this name.

To update one of the name, you also need to provide the row number in which the name to be replaced appear or its ott id.

See also

Examples

  
# NOT RUN {
   matched_names <- tnrs_match_names(c("holothuria", "diadema", "boletus"))
   inspect(matched_names, taxon_name="diadema")
   new_matched_names <- update(matched_names, taxon_name="diadema",
                               new_ott_id = 631176)
   new_matched_names

# }