List of studies used in the Tree of Life

Retrieve the detailed information for the list of studies used in the Tree of Life.

source_list(tax, ...)

# S3 method for tol_summary
source_list(tax, ...)

Arguments

tax
a list containing a source_id_map slot.

...
additional arguments (currently unused)

Value

a data frame

Details

This function takes the object resulting from tol_about(study_list = TRUE), tol_mrca(), tol_node_info(), and returns a data frame listing the tree_id, study_id and git_sha for the studies currently included in the Tree of Life.