Taxonomic MRCA

Taxonomic Least Inclusive Common Ancestor (MRCA)

taxonomy_mrca(ott_ids = NULL, ...)

# S3 method for taxon_mrca
tax_rank(tax, ...)

# S3 method for taxon_mrca
tax_name(tax, ...)

# S3 method for taxon_mrca
ott_id(tax, ...)

# S3 method for taxon_mrca
unique_name(tax, ...)

# S3 method for taxon_mrca
tax_sources(tax, ...)

# S3 method for taxon_mrca
flags(tax, ...)

# S3 method for taxon_mrca
is_suppressed(tax, ...)

Arguments

ott_ids
a vector of ott ids for the taxa whose MRCA is to be found (numeric).

...
additional arguments to customize the API request (see rotl package documentation).

tax
an object generated by the taxonomy_mrca function

Value

taxonomy_mrca returns a list about the taxonomic information relating to the MRCA for the ott_ids provided.

tax_rank returns a character vector of the taxonomic rank for the MRCA.

tax_name returns a character vector the Open Tree Taxonomy name for the MRCA.

ott_id returns a numeric vector of the ott id for the MRCA.

Details

Given a set of OTT ids, get the taxon that is the most recent common ancestor (the MRCA) of all the identified taxa.

Examples

# NOT RUN {
req <- taxonomy_mrca(ott_ids=c(515698,590452,643717))
tax_rank(req)
tax_name(req)
ott_id(req)
# }