Taxon information

Information about taxa.

Usage

taxonomy_taxon_info(
  ott_ids,
  include_children = FALSE,
  include_lineage = FALSE,
  include_terminal_descendants = FALSE,
  ...
)

# S3 method for taxon_info
tax_rank(tax, ...)

# S3 method for taxon_info
tax_name(tax, ...)

# S3 method for taxon_info
unique_name(tax, ...)

# S3 method for taxon_info
synonyms(tax, ...)

# S3 method for taxon_info
ott_id(tax, ...)

# S3 method for taxon_info
tax_sources(tax, ...)

# S3 method for taxon_info
is_suppressed(tax, ...)

# S3 method for taxon_info
flags(tax, ...)

Arguments

ott_ids

the ott ids of the taxon of interest (numeric or character containing only numbers)

include_children

whether to include information about all the children of this taxon. Default FALSE.

include_lineage

whether to include information about all the higher level taxa that include the ott_ids. Default FALSE.

include_terminal_descendants

whether to include the list of terminal ott_ids contained in the ott_ids provided.

...

additional arguments to customize the API request (see rotl package documentation).

tax

an object generated by the taxonomy_taxon_info function

Value

taxonomy_taxon_info returns a list detailing information about the taxa. tax_rank and

tax_name return a vector. synonyms returns a list whose elements are the synonyms for each of the

ott_id requested.

Details

Given a vector of ott ids, taxonomy_taxon_info returns information about the specified taxa.

The functions tax_rank, tax_name, and synonyms can extract this information from an object created by the taxonomy_taxon_info().

See also

Examples

if (FALSE) {
req <- taxonomy_taxon_info(ott_id=515698)
tax_rank(req)
tax_name(req)
synonyms(req)
}